+++ core/src/cn/harryh/arkpets/rpc/KWinPluginInterface.java
@@ -1,4 +1,4 @@
-package cn.harryh.arkpets.natives;
+package cn.harryh.arkpets.rpc;
 
 import org.freedesktop.dbus.TypeRef;
 import org.freedesktop.dbus.annotations.DBusBoundProperty;
